% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dates.R
\name{add_date_col}
\alias{add_date_col}
\title{Add a new date column that is floored by
a unit. The new date col's name will be
the first two letters of the unit.}
\usage{
add_date_col(df, col, unit)
}
\arguments{
\item{df}{Data frame.}

\item{col}{Column name of Date or POSIXct type.}

\item{unit}{Units as in \code{lubridate::floor_date()}.}
}
\value{
Data frame with a new floored date column.
}
\description{
Add a new date column that is floored by
a unit. The new date col's name will be
the first two letters of the unit.
}
