% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_tools.R
\name{check_complete_PDG}
\alias{check_complete_PDG}
\title{Check if a single PDG is complete (AMR and Cluster metadata)}
\usage{
check_complete_PDG(organism, PDG)
}
\arguments{
\item{organism}{a string ie 'Salmonella' or 'Campylobacter' etc}

\item{PDG}{}
}
\value{
TRUE/FALSE depending on if all URLs for download exist
}
\description{
Check if a single PDG is complete (AMR and Cluster metadata)
}
\examples{
#check_complete_PDG(URL)
}
