% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pangenome_tools.R
\name{cluster_genomes}
\alias{cluster_genomes}
\title{Cluster genomes at 4 levels from a pangenome gene PA matrix
Needs parallelDist, igraph,
genomes as rows and genes as columns?}
\usage{
cluster_genomes(
  dat_mat,
  pcut = 0,
  scut = 0.5,
  tcut = 0.75,
  qcut = 0.75,
  DIST_METHOD = "simpson",
  output_directory = NULL,
  write_dist = TRUE
)
}
\arguments{
\item{write_dist}{}
}
\value{

}
\description{
Cluster genomes at 4 levels from a pangenome gene PA matrix
Needs parallelDist, igraph,
genomes as rows and genes as columns?
}
