% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{country_vector}
\alias{country_vector}
\title{A named vector of country names}
\format{
A named vector of regex patterns (only US and UK have special entries)
}
\source{
the maps package
}
\usage{
country_vector
}
\description{
Used to help extract country of isolation from geo_loc_tag column,
so far only US and UK have special entries.  Taken from the maps package and slightly modified.
}
\keyword{datasets}
