% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dates.R
\name{daily}
\alias{daily}
\title{Add a daily variable to group by.}
\usage{
daily(df, col)
}
\arguments{
\item{df}{Data frame.}

\item{col}{Column name of Date or POSIXct type.}
}
\value{
Data frame with a new column \code{da}.
}
\description{
Add a daily variable to group by.
}
