% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genome_download_tools.R
\name{download_files}
\alias{download_files}
\title{Download specified files from NCBI ftp site}
\usage{
download_files(data, type)
}
\arguments{
\item{data}{a dataframe with columns created by make_download_urls() and make_dest_paths()}

\item{type}{the type of files you want to download, one of: 'fna', 'gbff', 'gff', 'gtf', 'faa', 'cds'}
}
\value{
the results of attempting to download the specified files,
A dataframe with two added columns:
\enumerate{
\item return of download.file, should be 0
\item error column from purrr::safely(), should contain any error messages.
}
}
\description{
Download specified files from NCBI ftp site
}
\examples{
# download_data \%>\% download_files('fna')
}
