% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genome_download_tools.R
\name{download_gbk_assembly_summary}
\alias{download_gbk_assembly_summary}
\title{Convenience function to download the assembly_summary.txt file from genbank}
\usage{
download_gbk_assembly_summary(destfile, organism = NULL)
}
\arguments{
\item{destfile}{passed to download.file()'s destfile, path to store the downloaded file}

\item{organism}{limit the gbk assembly summary to only this organism}
}
\value{
result of download.file
}
\description{
Convenience function to download the assembly_summary.txt file from genbank
}
\examples{
#not run download_gbk_assembly_summary(destfile='assembly_summary.txt')
}
