% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genome_download_tools.R
\name{download_genomes}
\alias{download_genomes}
\title{Download a set of specified genome files}
\usage{
download_genomes(data, type, PARALLEL = FALSE)
}
\arguments{
\item{data}{dataframe that both make_download_urls and make_dest_paths have run on}

\item{type}{one of the accepted types of files to download}

\item{PARALLEL}{should downloads be run in parallel, requires you to set your
future::plan()}
}
\value{
the original dataframe with columns added to represent the result of the download attempt
}
\description{
Download a set of specified genome files
}
\examples{
#
}
