% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_tools.R
\name{download_most_recent_complete}
\alias{download_most_recent_complete}
\title{Download the most recent complete metadata for a specified organism}
\usage{
download_most_recent_complete(organism, folder_prefix = NULL)
}
\arguments{
\item{organism}{a string ie 'Salmonella' or 'Campylobacter' etc}

\item{folder_prefix}{a string to append to the download path, ie './data/'}
}
\value{

}
\description{
Download the most recent complete metadata for a specified organism
}
\examples{
#download_most_recent_complete('Salmonella')
}
