% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_tools.R
\name{download_PDD_metadata}
\alias{download_PDD_metadata}
\title{Download Pathogen Detection metadata for a given organism}
\usage{
download_PDD_metadata(organism, PDG, folder_prefix = NULL)
}
\arguments{
\item{organism}{a string ie 'Salmonella' or 'Campylobacter' etc}

\item{PDG}{The Pathogen Detection accession number to download}

\item{folder_prefix}{a string to append to the download path, ie './data/'}
}
\value{
returns nothing, but probably should. Will initiate a download of the specified files
}
\description{
Download Pathogen Detection metadata for a given organism
}
\examples{
#download_PDD_metadata(organism = 'Campylobacter',PDG = 'PDG000000003.1517')
}
