% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_download_tools.R
\name{download_SNP_trees}
\alias{download_SNP_trees}
\title{Download SNP trees from a dataframe containing SNP_tree_urls and dests}
\usage{
download_SNP_trees(data)
}
\arguments{
\item{data}{dataframe with SNP tree URLs and dests}
}
\value{
the input dataframe with an added column indicating the status of the download
}
\description{
Download SNP trees from a dataframe containing SNP_tree_urls and dests
}
\examples{
# soon
}
