% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_pangenome_matrix}
\alias{example_pangenome_matrix}
\title{A gene presence absence matrix for a synthetic pangenome}
\format{
A presence/absence matrix with genes as rows and genomes as columns
}
\source{
pdtools:::generate_pangenome()
}
\usage{
example_pangenome_matrix
}
\description{
Used for pangenome function examples
}
\keyword{datasets}
