% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_clean.R
\name{extract_country}
\alias{extract_country}
\title{Extract a standardized country name from geo_loc_tag column}
\usage{
extract_country(meta, parallel = FALSE)
}
\arguments{
\item{meta}{A PDD metadata table}

\item{parallel}{boolean, should furrr be used to parallelize? need to set your future::plan()}
}
\value{
a two column tibble with target_acc and country as columns
}
\description{
Extract a standardized country name from geo_loc_tag column
}
\examples{
klebsiella_example_dat \%>\% extract_country()
}
