% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_cleanup_tools.R
\name{extract_state}
\alias{extract_state}
\title{Extract the state name from the geo_loc_name column, this only makes sence for USA isolates}
\usage{
extract_state(data)
}
\arguments{
\item{data}{a dataframe containing a \code{geo_loc_name} column}
}
\value{
a tibble with 2 columns, \code{target_acc} and \code{State}
}
\description{
Extract the state name from the geo_loc_name column, this only makes sence for USA isolates
}
\examples{
klebsiella_example_dat \%>\% extract_state()
}
