% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PDD_download.R
\name{find_most_recent_complete}
\alias{find_most_recent_complete}
\title{Find most recent complete PDG}
\usage{
find_most_recent_complete(organism)
}
\arguments{
\item{organism}{a string ie 'Salmonella' or 'Campylobacter' etc}
}
\value{

}
\description{
Find most recent complete PDG
}
\examples{
#find_most_recent_complete('Salmonella')
}
