% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dates.R
\name{floor_by}
\alias{floor_by}
\title{Floor date by days, weeks, months, etc.}
\usage{
floor_by(d, unit, week_start = 1)
}
\arguments{
\item{d}{Date or POSIXct vector}

\item{unit}{Units as in \code{lubridate::floor_date()}.}
}
\value{
Date vector floored by a specified unit.
}
\description{
Floor date by days, weeks, months, etc.
}
