% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_tools.R
\name{ftp_paths_from_assem_sum}
\alias{ftp_paths_from_assem_sum}
\title{Generate a two column tibble mapping ftp paths to assembly accessions}
\usage{
ftp_paths_from_assem_sum(filename)
}
\arguments{
\item{filename}{path to the ncbi genbank assembly_summary.txt
curl -O https://ftp.ncbi.nlm.nih.gov/genomes/genbank/bacteria/assembly_summary.txt}
}
\value{
returns a two column tibble of asm_acc and ftp
}
\description{
Generate a two column tibble mapping ftp paths to assembly accessions
}
\examples{
#ftp_paths <- ftp_paths_from_assem_sum('./data/assembly_summary.txt')
}
