% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pangenome_tools.R
\name{generate_genome_vector}
\alias{generate_genome_vector}
\title{Generate a genome tibble containing presence/absence of genes in a fake pangenome}
\usage{
generate_genome_vector(genome_name, num_genes, core_genome_fraction = 0.75)
}
\arguments{
\item{genome_name}{name of the genome to generate}

\item{num_genes}{number of genes the genome should contain}

\item{core_genome_fraction}{fraction of the genome that is part of the 'core pangenome'}
}
\value{
3 column tibble 1) genome_name; 2) gene_name 3) gene_presence
}
\description{
Generate a genome tibble containing presence/absence of genes in a fake pangenome
}
\examples{
#soon
}
