% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pangenome_tools.R
\name{generate_pangenome}
\alias{generate_pangenome}
\title{Generate a random synthetic pangenome gene presence absence matrix}
\usage{
generate_pangenome(
  num_genomes = 100,
  num_genes = 1000,
  core_genome_fraction = 0.75
)
}
\arguments{
\item{num_genomes}{number of genomes the matrix will contain}

\item{num_genes}{number of genes the matrix will contain}

\item{core_genome_fraction}{fraction of genes that are part of the core genome}
}
\value{
gene presence absence matrix (0/1), rows are genes, columns are genomes
}
\description{
Generate a random synthetic pangenome gene presence absence matrix
}
\examples{
generate_pangenome()
}
