% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_clean.R
\name{get_earliest_year}
\alias{get_earliest_year}
\title{add a Year column to PDD metadata containing the earliest year from the
available 'date' fields.}
\usage{
get_earliest_year(PDD_metadata_table)
}
\arguments{
\item{PDD_metadata_table}{an ncbi pathogen detection metadata table}
}
\value{
returns the input metadata table with an added \code{Year} column
}
\description{
add a Year column to PDD metadata containing the earliest year from the
available 'date' fields.
}
\examples{
get_earliest_year(klebsiella_example_dat)
}
