% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pangenome_select_representative_set.R
\name{get_gene_content_reps}
\alias{get_gene_content_reps}
\title{look for small sets of pangenomes that produce the desired gene content coverage of a pangenome}
\usage{
get_gene_content_reps(
  pan_mat,
  desired_coverage = 0.95,
  starting_set_size = 1,
  num_iters_per_size = 500,
  set_size_step = 1,
  PARALLEL = TRUE
)
}
\arguments{
\item{pan_mat}{input pangenome PA matrix}

\item{desired_coverage}{proportion of the pangenome to cover (.95)}

\item{starting_set_size}{initial number of genomes to use as a set size}

\item{num_iters_per_size}{number of samples to take for each set size setp}
}
\value{
tibble of sets of genomes that meet the desired coverage threshold, will be at least one set, but maybe more.
}
\description{
look for small sets of pangenomes that produce the desired gene content coverage of a pangenome
}
