% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_organism_table.R
\name{get_organism_table}
\alias{get_organism_table}
\title{Load the organism summary table}
\usage{
get_organism_table(
  browser = "firefox",
  url = "https://www.ncbi.nlm.nih.gov/pathogens/organisms/"
)
}
\arguments{
\item{browser}{'string designating the browser to use, default is firefox'}

\item{url}{'url to extract the table from.  default is https://www.ncbi.nlm.nih.gov/pathogens/organisms/}
}
\value{
a dataframe containing the organism information
}
\description{
Loads a table that lists the organisms in the NCBI pathogens project and the number of isolates
associated with each
}
\details{
Requires the RSelenium and janitor packages.
Rselenium was a bit of a pain for me to get working with firefox in my linux env
}
\examples{
'get_organisms_table()'
}
