% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pangenome_tools.R
\name{get_pangenome_representatives2}
\alias{get_pangenome_representatives2}
\title{Get pangenome representatives from a gene_vec_tibble}
\usage{
get_pangenome_representatives2(
  gene_vec_tibble,
  desired_coverage = 0.95,
  SEED = 3,
  best_possible_score = NULL,
  max_genomes = 1000
)
}
\arguments{
\item{gene_vec_tibble}{an object returned by pan_mat_to_gene_vec_tibble()}

\item{desired_coverage}{proportion of gene content desired 0-1}

\item{SEED}{random seed to use (for selecting 1st genome)}

\item{best_possible_score}{to save time you can pre-calculate the best possible score (total gene content of pangenome)}

\item{max_genomes}{Maximum number of genomes to select}
}
\value{
a list of 3; list(cumulative_genomes, scores, proportion_coverages)
}
\description{
Get pangenome representatives from a gene_vec_tibble
}
\examples{
generate_pangenome() \%>\% pan_mat_to_gene_vec_tibble() \%>\% get_pangenome_representatives2()
}
