% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pangenome_tools.R
\name{get_pangenome_representatives_jaccard}
\alias{get_pangenome_representatives_jaccard}
\title{Get a subset of a pangenome that tries to represent the gene presence absence diversity
Selects a random genome,
then selects the most distant genome from the selected genome
then}
\usage{
get_pangenome_representatives_jaccard(
  pan_mat = NULL,
  pan_dist = NULL,
  SEED = 3,
  verbose = FALSE,
  CUTOFF = 0.5,
  max_genomes = 1000
)
}
\arguments{
\item{pan_mat}{gene presence absence matrix}

\item{pan_dist}{distance matrix if not providing the PA matrix}

\item{SEED}{random seed}

\item{verbose}{include print statemetns?}

\item{CUTOFF}{stop choosing new genomes when distances are below this level}

\item{max_genomes}{stop choosing new genomes when this many genomes are chosen}
}
\value{
returns a tibble, 1) asm_acc, 2) min_jacc, 3) order
}
\description{
Get a subset of a pangenome that tries to represent the gene presence absence diversity
Selects a random genome,
then selects the most distant genome from the selected genome
then
}
\examples{
# get_pangenome_representatives_jaccard(pan_dist)
}
