% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_download_tools.R
\name{get_PDG_version}
\alias{get_PDG_version}
\title{get the version of the PDD metadata}
\usage{
get_PDG_version(data_dir)
}
\arguments{
\item{data_dir}{directory that contains the downloaded PDD metadata}
}
\value{
character vector length 1, PDG accession
}
\description{
get the version of the PDD metadata
}
