% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dates.R
\name{hourly}
\alias{hourly}
\title{Add an hourly variable to group by.}
\usage{
hourly(df, col)
}
\arguments{
\item{df}{Data frame.}

\item{col}{Column name of Date or POSIXct type.}
}
\value{
Data frame with a new column \code{da}.
}
\description{
Add an hourly variable to group by.
}
