% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_tools.R
\name{list_organisms}
\alias{list_organisms}
\title{List organisms available in the Pathogens database}
\usage{
list_organisms()
}
\value{
a tibble of available organisms
}
\description{
List organisms available in the Pathogens database
}
\examples{
 list_organisms()
}
