% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_tools.R
\name{list_PDGs}
\alias{list_PDGs}
\title{List available PDG accessions for an organism}
\usage{
list_PDGs(organism)
}
\arguments{
\item{organism}{a string ie 'Salmonella' or 'Campylobacter' etc}
}
\value{
returns a tibble or PDG accessions and release dates
}
\description{
List available PDG accessions for an organism
}
\examples{
 list_PDGs('Klebsiella')
}
