% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genome_download_tools.R
\name{make_dest_paths}
\alias{make_dest_paths}
\title{Make download destination paths}
\usage{
make_dest_paths(data, type, dest_dir)
}
\arguments{
\item{data}{A dataframe containing an asm_acc column}

\item{type}{the type of files you want to download, one of: 'fna', 'gbff', 'gff', 'gtf', 'faa', 'cds'}

\item{dest_dir}{path to the directory you want to use, must exist, should include a trailing '/'}
}
\value{
returns a dataframe with an added "{type}_dest" column containing the paths to pass to download.file
}
\description{
Make download destination paths
}
\examples{
# download_data \%>\% make_dest_faths(type='fna', dest_dir='./data/')
}
