% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_tools.R
\name{make_download_urls}
\alias{make_download_urls}
\title{generate ftp download paths for a selection of assembly accessions}
\usage{
make_download_urls(asm_acc, ftp_paths, type)
}
\arguments{
\item{asm_acc}{a vector of assembly accessions, in the same order as the ftp_paths}

\item{ftp_paths}{a vector of ftp_paths, in the same order as the asm_acc vector}

\item{type}{type of download path to generate, one of: 'fna', 'gbff', 'gff', 'gtf', 'faa', 'cds'}
}
\value{
returns a two column tibble, 1st column= asm_acc ; 2nd column= download_path
}
\description{
generate ftp download paths for a selection of assembly accessions
}
\examples{
make_download_urls(asm_acc = klebsiella_example_dat$asm_acc,
                             ftp_paths = klebsiella_example_dat$ftp_path, type='fna')

}
