% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_tools.R
\name{make_fna_urls}
\alias{make_fna_urls}
\title{generate fasta ftp download paths for a vector of assembly accessions}
\usage{
make_fna_urls(asm_accessions, assembly_summary_path)
}
\arguments{
\item{asm_accessions}{vector of assembly accessions to generate ftp paths for}

\item{assembly_summary_path}{path to an assembly_summary.txt file downloaded from ncbi}
}
\value{
returns a vector of ftp paths for use with curl wget etc.
}
\description{
generate fasta ftp download paths for a vector of assembly accessions
}
\examples{
# tst <- make_fna_urls(klebsiella_example_dat$asm_acc, './data/assembly_summary.txt')
}
