% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_tools.R
\name{make_ftp_paths}
\alias{make_ftp_paths}
\title{generate ftp site paths for a selection of assembly accessions}
\usage{
make_ftp_paths(data, assembly_summary_path)
}
\arguments{
\item{data}{a dataframe containing an asm_acc}

\item{assembly_summary_path}{path to genbank assembly_summary.txt, see download_gbk_assembly_summary()}
}
\value{
a two column tibble 1= asm_acc ; 2= ftp_path
}
\description{
generate ftp site paths for a selection of assembly accessions
}
\examples{
#make_ftp_paths(klebsiella_example_data './test/assembly_summary.txt')
}
