% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_download_tools.R
\name{make_SNP_tree_dest}
\alias{make_SNP_tree_dest}
\title{Make SNP tree download destination paths}
\usage{
make_SNP_tree_dest(data, data_dir)
}
\arguments{
\item{data}{A dataframe containing the download paths for desired SNP trees}

\item{data_dir}{a path to a directory to store the downloaded trees in}
}
\value{
the input dataframe with an added column containing the SNP_tree_dest column
}
\description{
Make SNP tree download destination paths
}
\examples{
# soon
}
