% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_tools.R
\name{make_SNPtree_urls}
\alias{make_SNPtree_urls}
\title{generate ftp site download urls for all SNP trees containing the provided isolates}
\usage{
make_SNPtree_urls(organism, data, PDG)
}
\arguments{
\item{organism}{a string ie 'Salmonella' or 'Campylobacter' etc}

\item{data}{a metadata table, must contain the column 'PDS_acc' from merging in the SNP cluster data}

\item{PDG}{The PDG version the metadata is from.}
}
\value{
returns a vector of ftp download urls for each tar.gz file containing the SNP tree info
}
\description{
generate ftp site download urls for all SNP trees containing the provided isolates
}
\examples{
make_SNPtree_urls(organism = 'Klebsiella',
 data = klebsiella_example_dat, PDG = 'PDG000000012.1053')
}
