% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pangenome_tools.R
\name{mark_outliers}
\alias{mark_outliers}
\title{Mark outliers from a distance matrix}
\usage{
mark_outliers(DIST, outlier_prob = 0.99)
}
\arguments{
\item{DIST}{a dist object}

\item{outlier_prob}{probability to consider an entity an outlier}
}
\value{
a tibble with an outlier designation for each genome
}
\description{
Mark outliers from a distance matrix
}
\examples{
mark_outliers(example_pan_dist, outlier_prob=.95)
}
