% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dates.R
\name{monthly}
\alias{monthly}
\title{Add a monthly variable to group by.}
\usage{
monthly(df, col)
}
\arguments{
\item{df}{Data frame.}

\item{col}{Column name of Date or POSIXct type.}
}
\value{
Data frame with a new column \code{mo}.
}
\description{
Add a monthly variable to group by.
}
