% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pangenome_tools.R
\name{pan_mat_to_gene_vec_tibble}
\alias{pan_mat_to_gene_vec_tibble}
\title{Convert a pangenome PA matrix to a tibble of gene vectors}
\usage{
pan_mat_to_gene_vec_tibble(pan_mat)
}
\arguments{
\item{pan_mat}{presence absence matrix, genes are columns, rows are genomes 1/0}
}
\value{
a tibble with gene vectors for each genome
}
\description{
Convert a pangenome PA matrix to a tibble of gene vectors
}
\examples{
#soon
}
