% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pangenome_tools.R
\name{remove_strict_core}
\alias{remove_strict_core}
\title{Removes genes present in all genomes from pangenome presence/absence matrix}
\usage{
remove_strict_core(pan_PA, rows_are_genes = NULL)
}
\arguments{
\item{pan_PA}{a pangenome presence absence matrix}

\item{rows_are_genes}{a logical indicating if genes are rows in the matrix}
}
\value{
returns a pangenome presence/absence matrix with the strict core removed
}
\description{
Removes genes present in all genomes from pangenome presence/absence matrix
}
\examples{
generate_pangenome() \%>\% remove_strict_core()
}
