% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pangenome_tools.R
\name{res_plot_dat}
\alias{res_plot_dat}
\title{Generate a dataframe for plotting the progress of selecting a set of genomes}
\usage{
res_plot_dat(result, seed_num)
}
\arguments{
\item{seed_num}{}
}
\value{

}
\description{
Generate a dataframe for plotting the progress of selecting a set of genomes
}
