% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_clean.R
\name{return_ag_match}
\alias{return_ag_match}
\title{Return a string of names of the matching regex patterns}
\usage{
return_ag_match(pattern_vec, search_string)
}
\arguments{
\item{pattern_vec}{a named vector of regex patterns to use}

\item{search_string}{a sting to search for matches}
}
\value{
names of matching patterns concatenated with '_'
}
\description{
Return a string of names of the matching regex patterns
}
\examples{
#pat_vec <- c(Swine='hog|swine|sow'); pdtools:::return_ag_match(pattern_vec=pat_vec, 'Hog')

}
