% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_clean.R
\name{return_earliest_year}
\alias{return_earliest_year}
\title{return a Year column containing the earliest year from the
available 'date' fields.}
\usage{
return_earliest_year(PDD_metadata_table)
}
\arguments{
\item{PDD_metadata_table}{an ncbi pathogen detection metadata table}
}
\value{
returns the a tibble with a \code{Year} column,
}
\description{
return a Year column containing the earliest year from the
available 'date' fields.
}
\examples{
return_earliest_year(klebsiella_example_dat)
}
