% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pangenome_tools.R
\name{return_set_score}
\alias{return_set_score}
\title{return the coverage score (proportion coverage) of a random set of genomes}
\usage{
return_set_score(pan_mat, set_size)
}
\arguments{
\item{pan_mat}{input pangenome PA matrix}

\item{set_size}{number of genomes to randomly select}
}
\value{
returns a named list of length 2, [\link{1}] = scores, [\link{2}] genome indicies
}
\description{
return the coverage score (proportion coverage) of a random set of genomes
}
\examples{
#soon
}
