% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genome_download_tools.R
\name{supported_download_types}
\alias{supported_download_types}
\title{helper function to check if user supplied type is in the supported types}
\usage{
supported_download_types(type)
}
\arguments{
\item{type}{a user input string to check}
}
\value{

}
\description{
helper function to check if user supplied type is in the supported types
}
\examples{
# supported_download_types('fna')
}
