% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dates.R
\name{weekly}
\alias{weekly}
\title{Add a weekly variable to group by.}
\usage{
weekly(df, col)
}
\arguments{
\item{df}{Data frame.}

\item{col}{Column name of Date or POSIXct type.}
}
\value{
Data frame with a new column \code{we}.
}
\description{
Add a weekly variable to group by.
}
