% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dates.R
\name{yearly}
\alias{yearly}
\title{Add a yearly variable to group by.}
\usage{
yearly(df, col)
}
\arguments{
\item{df}{Data frame.}

\item{col}{Column name of Date or POSIXct type.}
}
\value{
Data frame with a new column \code{ye}.
}
\description{
Add a yearly variable to group by.
}
