% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pangenome_tools.R
\name{build_ppanggolin_file_fastas}
\alias{build_ppanggolin_file_fastas}
\title{Build a ppanggolin file from fastas}
\usage{
build_ppanggolin_file_fastas(
  complete_genome_paths = NULL,
  incomplete_genome_paths = NULL
)
}
\arguments{
\item{incomplete_genome_paths}{}
}
\value{
a tibble satisfying the ppanggolin file requirements with the complete genome contigs indicated as circular
}
\description{
Build a ppanggolin file from fastas
}
\examples{
#soon
}
