% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pangenome_tools.R
\name{calculate_novelty}
\alias{calculate_novelty}
\title{Calculate genome novelty from a list of selection sets}
\usage{
calculate_novelty(selection_set_results)
}
\arguments{
\item{selection_set_results}{}
}
\value{
a dataframe with the novelty scores of all genomes in the list
}
\description{
Calculate genome novelty from a list of selection sets
}
\examples{
# soon
}
