% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PDD_download.R
\name{check_complete_PDG}
\alias{check_complete_PDG}
\title{Check if a single PDG is complete (AMR and Cluster metadata)}
\usage{
check_complete_PDG(url)
}
\arguments{
\item{url}{}
}
\value{

}
\description{
Check if a single PDG is complete (AMR and Cluster metadata)
}
\examples{
check_complete_PDG(URL)
}
