% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pangenome_tools.R
\name{cluster_genomes}
\alias{cluster_genomes}
\title{Cluster genomes at 4 levels from a pangenome gene PA matrix
Needs parallelDist, igraph,
genomes as rows and genes as columns?}
\usage{
cluster_genomes(
  dat_mat,
  pcut = 0,
  scut = 0.5,
  tcut = 0.75,
  qcut = 0.75,
  DIST_METHOD = "binary",
  output_directory = NULL,
  write_dist = TRUE,
  write_graph = TRUE
)
}
\arguments{
\item{dat_mat}{pan genome presence absence matrix (rows are genes)}

\item{pcut}{weight for graph pruning at 1st level}

\item{scut}{weight for graph pruning at 2nd level}

\item{tcut}{weight for graph pruning at 3rd level}

\item{qcut}{weight for graph pruning at 4th level}

\item{DIST_METHOD}{distance method for building graph (passed to parallelDist())}

\item{output_directory}{output directory to save the graph in}

\item{write_dist}{logical, should the dist object be written?}

\item{write_graph}{logical, should the graph be written to an rds file?}
}
\value{
returns a tibble with a cluster designation for each genome at 4 levels
}
\description{
Cluster genomes at 4 levels from a pangenome gene PA matrix
Needs parallelDist, igraph,
genomes as rows and genes as columns?
}
\examples{
# note the t() because the dist() function finds dists between rows of a matrix
cluster_genomes(dat_mat=t(example_pangenome_matrix),
                pcut=.80,
                scut=.85,
                tcut=.90,
                qcut=.95,
                write_dist=FALSE,
                write_graph=FALSE)



}
