% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PDD_download.R
\name{download_PDD_metadata}
\alias{download_PDD_metadata}
\title{Download Pathogen Detection metadata for a given organism}
\usage{
download_PDD_metadata(organism, PDG, folder_prefix = NULL)
}
\arguments{
\item{organism}{a string ie 'Salmonella' or 'Campylobacter' etc}

\item{folder_prefix}{}
}
\value{

}
\description{
Download Pathogen Detection metadata for a given organism
}
