% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_clean.R
\name{extract_collection_agency}
\alias{extract_collection_agency}
\title{extract collecting agency from several metadata fields}
\usage{
extract_collection_agency(meta, parallel = FALSE)
}
\arguments{
\item{meta}{an ncbi pathogen detection metadata table}

\item{parallel}{boolean, should furrr be used to parallelize? need to set your future::plan()}
}
\value{
a tibble with 2 columns, 1 = target_acc, 2 = collection_agency
}
\description{
extract collecting agency from several metadata fields
}
\examples{
klebsiella_example_dat \%>\% extract_collection_agency()
}
