% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pangenome_tools.R
\name{get_pangenome_representatives}
\alias{get_pangenome_representatives}
\title{Select a minimum set of genomes that best represent the gene content of the pangenome}
\usage{
get_pangenome_representatives(pan_mat, desired_coverage = 0.95, SEED = 3)
}
\arguments{
\item{SEED}{}
}
\value{

}
\description{
Select a minimum set of genomes that best represent the gene content of the pangenome
}
