% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{klebsiella_example_dat}
\alias{klebsiella_example_dat}
\title{NCBI Pathogen Detection metadata for 200 Klebsiella isolates}
\format{
A data frame with 200 rows and 64 variables:
}
\source{
\url{https://ftp.ncbi.nlm.nih.gov/pathogen/Results/Klebsiella/}
}
\usage{
klebsiella_example_dat
}
\description{
An example dataset of 100 Klebsiella isolates included for testing and
example purposes
}
\keyword{datasets}
