% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_tools.R
\name{make_download_urls}
\alias{make_download_urls}
\title{make specific ftp download paths for a dataframe with ftp_paths and assembly accessions}
\usage{
make_download_urls(data, type)
}
\arguments{
\item{data}{a dataframe with the columns 'ftp_path' and 'asm_acc'
ftp_path should be a column produced by the function make_ftp_paths()}

\item{type}{type of download path to generate, one of: 'fna', 'gbff', 'gff', 'gtf', 'faa', 'cds'}
}
\value{
returns the original dataframe with an added column, named "{type}_download"
}
\description{
make specific ftp download paths for a dataframe with ftp_paths and assembly accessions
}
\examples{
make_download_urls(klebsiella_example_dat, type='fna')

}
