% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PDD_download.R
\name{make_fna_urls}
\alias{make_fna_urls}
\title{Generate ftp urls for fna files from a vector of genbank assembly accessions}
\usage{
make_fna_urls(asm_accessions, assembly_summary)
}
\arguments{
\item{asm_accessions}{character vector of genbank accessions}

\item{assembly_summary}{assembly summary downloaded from ncbi, must have asm_acc column}
}
\value{
returns a vector of ftp urls
}
\description{
Generate ftp urls for fna files from a vector of genbank assembly accessions
}
