% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pangenome_tools.R
\name{pick_derep_sets}
\alias{pick_derep_sets}
\title{Pick multiple de-replication sets from a pangenome, uses furrr::future_map
Be sure to set your 'plan()'!!!}
\usage{
pick_derep_sets(pan_PA, output_file, num_sets = 25, desired_coverage = 0.95)
}
\arguments{
\item{pan_PA}{A pangenome gene presence/absence matrix, genomes as columns,
genes as rows}

\item{output_file}{a path to save the resulting R object to (RDS format)}

\item{num_sets}{The number of sets to select (25)}

\item{desired_coverage}{the proportion of genes in the pangenome to cover (.99)}
}
\value{
a tibble with two columns:
\enumerate{
\item the random seeds used,
\item list column containing the dereplication sets
}
}
\description{
Pick multiple de-replication sets from a pangenome, uses furrr::future_map
Be sure to set your 'plan()'!!!
}
\examples{
# soon
}
