% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pangenome_tools.R
\name{res_plot_dat}
\alias{res_plot_dat}
\title{Generate a dataframe for plotting the progress of selecting a set of genomes}
\usage{
res_plot_dat(result, seed_num)
}
\arguments{
\item{result}{TODO}

\item{seed_num}{TODO}
}
\value{
returns a tibble containing the score of the set with each additional genome
}
\description{
Generate a dataframe for plotting the progress of selecting a set of genomes
}
\examples{
# res_plot_dat()
}
